Factor out prepare_commit_msg module. #refactor-into-modules
